Refactor Http Request
